requirements.txt
annotated-types==0.7.0
anyio==4.4.0
autoflake==2.3.1
black==24.8.0
cfgv==3.4.0
click==8.1.7
distlib==0.3.8
exceptiongroup==1.2.2
fastapi==0.112.1
filelock==3.15.4
flake8==7.1.1
h11==0.14.0
identify==2.6.0
idna==3.7
iniconfig==2.0.0
isort==5.13.2
mccabe==0.7.0
mypy-extensions==1.0.0
mysql-connector-python==9.0.0
nodeenv==1.9.1
numpy==2.0.1
packaging==24.1
pandas==2.2.2
passlib[bcrypt]==1.7.4
pathspec==0.12.1
platformdirs==4.2.2
pluggy==1.5.0
pre-commit==3.8.0
pycodestyle==2.12.1
pydantic==2.8.2
pydantic_core==2.20.1
pyflakes==3.2.0
PyJWT==2.9.0
pytest==8.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.2
six==1.16.0
sniffio==1.3.1
starlette==0.38.2
tomli==2.0.1
typing_extensions==4.12.2
tzdata==2024.1
uvicorn==0.30.6
virtualenv==20.26.3